Use new tooltip in LockWithTooltip
